moved createdir to utils
