+++ win/ISLETrain/ISLETrain.vcxproj
@@ -35,6 +35,7 @@
   <ItemGroup>
     <ClCompile Include="..\..\drivers\ISLETrain.cpp" />
     <ClCompile Include="..\..\src\denseMatrix.cpp" />
+    <ClCompile Include="..\..\src\logger.cpp" />
     <ClCompile Include="..\..\src\sparseMatrix.cpp" />
     <ClCompile Include="..\..\src\trainer.cpp" />
     <ClCompile Include="..\..\src\utils.cpp" />
