+++ win/ISLETrain/ISLETrain.vcxproj.filters
@@ -62,5 +62,8 @@
     <ClCompile Include="..\..\src\sparseMatrix.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\logger.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
 </Project>
